feat: FileReader updated

+++ src/include/map.h
@@ -9,7 +9,7 @@
 typedef struct {
   uint16_t width;
   uint16_t height;
-  uint32_t buffer[AC_BUFFER_SIZE];
+  uint16_t buffer[AC_BUFFER_SIZE];
 } Map;
 
 #if defined(__cplusplus)
